refactor: #9 컴포넌트 폴더/파일명 네이밍 컨벤션 통일
- 폴더명 PascalCase → camelCase (Button → button, Header → header 등)
- types/variants/stories 파일명 소문자로 변경 (Button.types.ts → button.types.ts)
- 컴포넌트 파일(.tsx)은 PascalCase 유지
- 변경된 임포트 경로 수정 (app/layout.tsx, Header.tsx, Button.tsx)
- .gitignore에 Figma 임시 파일 추가
