+++ src/components/domain/fashionTasteCard/FashionTasteCard.tsx
