+++ src/components/domain/fashionTasteCard/index.ts
