--- src/components/layout/Header/index.ts
@@ -1,2 +1,0 @@
-export { Header } from "./Header";
-export type { IHeaderProps } from "./Header.types";
