+++ src/components/layout/header/index.ts
@@ -1,0 +1,2 @@
+export { Header } from "./Header";
+export type { IHeaderProps } from "./header.types";
