+++ src/components/result/recommendCard/RecommendCard.tsx
