+++ src/components/result/recommendCard/index.ts
