+++ src/components/result/shareCard/ShareCard.tsx
