+++ src/components/result/shareCard/index.ts
