+++ src/components/result/styleLabel/StyleLabel.tsx
