+++ src/components/result/styleLabel/index.ts
